+++ Test/SmallCheck/Property.hs
@@ -7,15 +7,13 @@
 --
 -- Properties and tools to construct them.
 --------------------------------------------------------------------
-{-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE MultiParamTypeClasses, FlexibleInstances, TypeFamilies,
+             ScopedTypeVariables #-}
 module Test.SmallCheck.Property (
   -- * Basic definitions
-  TestCase(..),
-  TestResult(..),
-  resultIsOk,
-
   Property, Depth, Testable(..),
-  property, mkProperty,
+  SC, Stats(..), Example,
+  property,
 
   -- * Constructing tests
   (==>), exists, existsDeeperBy, exists1, exists1DeeperBy,
@@ -30,63 +28,66 @@ module Test.SmallCheck.Property (
   ) where
 
 import Test.SmallCheck.Series
+import Test.SmallCheck.Monad
+import Control.Monad
+import Control.Monad.Logic
 import Data.Typeable
 
-data TestResult
-    = Pass
-    | Fail
-    | Inappropriate
-        -- ^ 'Inappropriate' means that the precondition of '==>'
-        -- was not satisfied
-data TestCase = TestCase { result :: TestResult, arguments :: [String] }
-
 -- | Wrapper type for 'Testable's
-newtype Property = Property (Depth -> [TestCase])
-  deriving Typeable
+newtype Property m = Property (SC m Example)
+
+instance Typeable1 m => Typeable (Property m)
+  where
+    typeOf _ =
+      mkTyConApp
+        (mkTyCon3 "smallcheck" "Test.SmallCheck.Property" "Property")
+        [typeOf (undefined :: m ())]
 
 -- | Wrap a 'Testable' into a 'Property'
-property :: Testable a => a -> Property
+property :: Testable m a => a -> Property m
 property = Property . test
 
--- | A lower-level way to create properties. Use 'property' if possible.
---
--- The argument is a function that produces the list of results given the depth
--- of testing.
-mkProperty :: (Depth -> [TestCase]) -> Property
-mkProperty = Property
-
 -- | Anything of a 'Testable' type can be regarded as a \"test\"
-class Testable a where
-  test :: a -> Depth -> [TestCase]
+class Monad m => Testable m a where
+  test :: a -> SC m Example
 
-instance Testable Bool where
-  test b _ = [TestCase (boolToResult b) []]
+instance Monad m => Testable m Bool where
+  test b = runTestHook >> record (boolToResult b)
 
-instance (Serial a, Show a, Testable b) => Testable (a->b) where
+instance (Serial m a, Show a, Testable m b) => Testable m (a->b) where
   test f = f' where Property f' = forAll series f
 
-instance Testable Property where
-  test (Property f) d = f d
+instance (Monad m, m ~ n) => Testable n (Property m) where
+  test (Property f) = f
 
-forAll :: (Show a, Testable b) => Series a -> (a->b) -> Property
-forAll xs f = Property $ \d ->
-  [ r{arguments = show x : arguments r}
-  | x <- xs d, r <- test (f x) d ]
+forAll :: (Show a, Testable m b) => Series m a -> (a->b) -> Property m
+forAll xs f = Property $ do
+  x <- xs
+  searchCounterexamples $
+    addArgument (show x) $
+    test (f x)
 
-forAllElem :: (Show a, Testable b) => [a] -> (a->b) -> Property
-forAllElem xs = forAll (const xs)
+forAllElem :: (Show a, Testable m b) => [a] -> (a->b) -> Property m
+forAllElem xs = forAll $ generate $ const xs
 
-existence :: (Show a, Testable b) => Bool -> Series a -> (a->b) -> Property
-existence u xs f = Property existenceDepth
-  where
-  existenceDepth d = [ TestCase (boolToResult valid) arguments ]
-    where
-    witnesses = [ show x | x <- xs d, all (resultIsOk . result) (test (f x) d) ]
-    valid     = enough witnesses
-    enough    = if u then unique else (not . null)
-    arguments = if valid then []
-                else if null witnesses then ["non-existence"]
-                else "non-uniqueness" : take 2 witnesses
+existence :: (Show a, Testable m b) => Bool -> Series m a -> (a->b) -> Property m
+existence u xs f = Property $ do
+  let
+    search = do
+      x <- xs
+      searchExamples $ addArgument (show x) $ test (f x)
+
+  first <- msplit search
+
+  case first of
+    Nothing -> return ["non-existence"]
+    Just (x1, search') | u -> do
+      second <- msplit search'
+      case second of
+        Nothing -> mzero
+        Just (x2, _) -> return $ concat [["non-uniqueness"], x1, x2]
+
+      | otherwise -> mzero
 
 unique :: [a] -> Bool
 unique [_] = True
@@ -103,26 +104,26 @@ resultIsOk r =
 boolToResult :: Bool -> TestResult
 boolToResult b = if b then Pass else Fail
 
-thereExists :: (Show a, Testable b) => Series a -> (a->b) -> Property
+thereExists :: (Show a, Testable m b) => Series m a -> (a->b) -> Property m
 thereExists = existence False
 
-thereExists1 :: (Show a, Testable b) => Series a -> (a->b) -> Property
+thereExists1 :: (Show a, Testable m b) => Series m a -> (a->b) -> Property m
 thereExists1 = existence True
 
-thereExistsElem :: (Show a, Testable b) => [a] -> (a->b) -> Property
-thereExistsElem xs = thereExists (const xs)
+thereExistsElem :: (Show a, Testable m b) => [a] -> (a->b) -> Property m
+thereExistsElem xs = thereExists $ generate $ const xs
 
-thereExists1Elem :: (Show a, Testable b) => [a] -> (a->b) -> Property
-thereExists1Elem xs = thereExists1 (const xs)
+thereExists1Elem :: (Show a, Testable m b) => [a] -> (a->b) -> Property m
+thereExists1Elem xs = thereExists1 $ generate $ const xs
 
 -- | @'exists' p@ holds iff it is possible to find an argument @a@ (within the
 -- depth constraints!) satisfying the predicate @p@
-exists :: (Show a, Serial a, Testable b) => (a->b) -> Property
+exists :: (Show a, Serial m a, Testable m b) => (a->b) -> Property m
 exists = thereExists series
 
 -- | Like 'exists', but additionally require the uniqueness of the
 -- argument satisfying the predicate
-exists1 :: (Show a, Serial a, Testable b) => (a->b) -> Property
+exists1 :: (Show a, Serial m a, Testable m b) => (a->b) -> Property m
 exists1 = thereExists1 series
 
 -- | The default testing of existentials is bounded by the same depth as their
@@ -134,20 +135,20 @@ exists1 = thereExists1 series
 -- interpretation of existential properties can make testing of a valid property
 -- fail at all depths. Here is a contrived but illustrative example:
 --
--- >prop_append1 :: [Bool] -> [Bool] -> Property
+-- >prop_append1 :: [Bool] -> [Bool] -> Property m
 -- >prop_append1 xs ys = exists $ \zs -> zs == xs++ys
 --
 -- 'existsDeeperBy' transforms the depth bound by a given @'Depth' -> 'Depth'@ function:
 --
--- >prop_append2 :: [Bool] -> [Bool] -> Property
+-- >prop_append2 :: [Bool] -> [Bool] -> Property m
 -- >prop_append2 xs ys = existsDeeperBy (*2) $ \zs -> zs == xs++ys
-existsDeeperBy :: (Show a, Serial a, Testable b) => (Depth->Depth) -> (a->b) -> Property
-existsDeeperBy f = thereExists (series . f)
+existsDeeperBy :: (Show a, Serial m a, Testable m b) => (Depth->Depth) -> (a->b) -> Property m
+existsDeeperBy f = thereExists $ localDepth f series
 
 -- | Like 'existsDeeperBy', but additionally require the uniqueness of the
 -- argument satisfying the predicate
-exists1DeeperBy :: (Show a, Serial a, Testable b) => (Depth->Depth) -> (a->b) -> Property
-exists1DeeperBy f = thereExists1 (series . f)
+exists1DeeperBy :: (Show a, Serial m a, Testable m b) => (Depth->Depth) -> (a->b) -> Property m
+exists1DeeperBy f = thereExists1 $ localDepth f series
 
 infixr 0 ==>
 
@@ -156,13 +157,13 @@ infixr 0 ==>
 -- testing a propositional-logic module (see examples/logical), we might
 -- define:
 --
--- >prop_tautEval :: Proposition -> Environment -> Property
+-- >prop_tautEval :: Proposition -> Environment -> Property m
 -- >prop_tautEval p e =
 -- >  tautology p ==> eval p e
 --
 -- But here is an alternative definition:
 --
--- >prop_tautEval :: Proposition -> Property
+-- >prop_tautEval :: Proposition -> Property m
 -- >prop_taut p =
 -- >  tautology p ==> \e -> eval p e
 --
@@ -172,8 +173,6 @@ infixr 0 ==>
 -- The second definition is far better as the test-space is
 -- reduced from PE to T'+TE where P, T, T' and E are the numbers of
 -- propositions, tautologies, non-tautologies and environments.
-(==>) :: Testable a => Bool -> a -> Property
+(==>) :: Testable m a => Bool -> a -> Property m
 True ==>  x = Property (test x)
-False ==> x = Property (const [nothing])
-    where
-    nothing = TestCase { result = Inappropriate, arguments = [] }
+False ==> x = Property $ runTestHook >> record Inappropriate